SDK_LEARNING/LEARNING_WITH_ONLINE_CLASS/PRACTICE_CODE/STREANMING/pyproject.toml
[project]
name = "practice-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.5.5",
    "openai-agents>=0.0.14",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "ruff>=0.11.8",
]
